backend/requirements.txt
certifi==2017.4.17
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
Django==1.11.2
django-filter==1.0.4
django-filters==0.2.1
django-floppyforms==1.7.0
django-rest-swagger==2.1.2
djangorestframework==3.6.3
djangorestframework-gis==0.11.2
idna==2.5
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
openapi-codec==1.3.1
psycopg2==2.7.1
pytz==2017.2
requests==2.18.1
simplejson==3.11.1
uritemplate==3.0.0
urllib3==1.21.1
